Add files via upload
This is a sample of 'Girl'. This song is about a romantic break up between 2 girls. With such a song I would like to highlight a subject that can be still sensitve nowadays and I hope that it can touch the heart of anyone who will be listening to it. Love is Love. It has no boundaries, no colour, no age, no gender.