Add files via upload
In this article dated from 2005, Vanuella explains the origin of her name. Indeed her name is a mix of 'Vanuatu', the country of her father, and 'Emmanuel: her father's name. Vanuella also talks about her passion for music and her passion for giving her personal interpretation of international songs. The article is also highlighting the fact that the songs of the artist have been played on all radios in New Caledonia.