Add files via upload
This short video was filmed during Vanuella's concert in October 2024 for the Telethon in New Caledonia. This is a small extract of Vanuella singing her hit 'Soleil' which made her famous in her country. The crowd is dancing and enjoying the song! Vanuella is performing with Cedric Lavigne, a famous percussionist in Noumea.