Add files via upload
In this article, Vanuella explains the origins of her name: 'Vanuella' is a mix of 'Vanuatu', the country of her father, and 'Emmanuel', his name! 
The artist is also talking about her passion for music and the article is highlighting the fact that her songs have been played in all radios in New Caledonia.