Add files via upload
'Another way' is a song wrtitten by Vanuella. The lyrics, although they are subtil, refer to the loss of a loved one. With this song, the singer shares with us the fact that she believes that even if we don't see those who have left us, still we feel their presence and we feel protected by them.